automation/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# API Clients
httpx==0.25.1
praw==7.7.1
tweepy==4.14.0
pytrends==4.9.2

# AI/ML
openai==1.3.5
transformers==4.35.0
torch==2.1.0

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
